+++ app/src/main/java/com/boardgamegeek/db/GameDao.kt
@@ -7,13 +7,14 @@ import android.content.ContentResolver
 import android.content.ContentValues
 import android.graphics.Color
 import android.net.Uri
-import com.boardgamegeek.*
+import com.boardgamegeek.BggApplication
 import com.boardgamegeek.entities.*
 import com.boardgamegeek.extensions.*
 import com.boardgamegeek.livedata.AbsentLiveData
 import com.boardgamegeek.livedata.RegisteredLiveData
 import com.boardgamegeek.provider.BggContract
 import com.boardgamegeek.provider.BggContract.*
+import com.boardgamegeek.provider.BggContract.Collection
 import com.boardgamegeek.provider.BggDatabase.*
 import com.boardgamegeek.util.DataUtils
 import com.boardgamegeek.util.NotificationUtils
@@ -187,15 +188,18 @@ class GameDao(private val context: BggApplication) {
         }
     }
 
-    fun loadDesigners(gameId: Int): LiveData<List<Pair<Int, String>>> {
+    fun loadDesigners(gameId: Int): LiveData<List<GameDetailEntity>> {
         if (gameId == BggContract.INVALID_ID) return AbsentLiveData.create()
         val uri = Games.buildDesignersUri(gameId)
         return RegisteredLiveData(context, uri) {
-            val results = arrayListOf<Pair<Int, String>>()
+            val results = arrayListOf<GameDetailEntity>()
             context.contentResolver.load(uri)?.use {
                 if (it.moveToFirst()) {
                     do {
-                        results.add(it.getInt(Designers.DESIGNER_ID) to it.getString(Designers.DESIGNER_NAME))
+                        results.add(GameDetailEntity(
+                                it.getInt(Designers.DESIGNER_ID),
+                                it.getString(Designers.DESIGNER_NAME)
+                        ))
                     } while (it.moveToNext())
                 }
             }
@@ -203,15 +207,18 @@ class GameDao(private val context: BggApplication) {
         }
     }
 
-    fun loadArtists(gameId: Int): LiveData<List<Pair<Int, String>>> {
+    fun loadArtists(gameId: Int): LiveData<List<GameDetailEntity>> {
         if (gameId == BggContract.INVALID_ID) return AbsentLiveData.create()
         val uri = Games.buildArtistsUri(gameId)
         return RegisteredLiveData(context, uri) {
-            val results = arrayListOf<Pair<Int, String>>()
+            val results = arrayListOf<GameDetailEntity>()
             context.contentResolver.load(uri)?.use {
                 if (it.moveToFirst()) {
                     do {
-                        results.add(it.getInt(Artists.ARTIST_ID) to it.getString(Artists.ARTIST_NAME))
+                        results.add(GameDetailEntity(
+                                it.getInt(Artists.ARTIST_ID),
+                                it.getString(Artists.ARTIST_NAME)
+                        ))
                     } while (it.moveToNext())
                 }
             }
@@ -219,15 +226,18 @@ class GameDao(private val context: BggApplication) {
         }
     }
 
-    fun loadPublishers(gameId: Int): LiveData<List<Pair<Int, String>>> {
+    fun loadPublishers(gameId: Int): LiveData<List<GameDetailEntity>> {
         if (gameId == BggContract.INVALID_ID) return AbsentLiveData.create()
         val uri = Games.buildPublishersUri(gameId)
         return RegisteredLiveData(context, uri) {
-            val results = arrayListOf<Pair<Int, String>>()
+            val results = arrayListOf<GameDetailEntity>()
             context.contentResolver.load(uri)?.use {
                 if (it.moveToFirst()) {
                     do {
-                        results.add(it.getInt(Publishers.PUBLISHER_ID) to it.getString(Publishers.PUBLISHER_NAME))
+                        results.add(GameDetailEntity(
+                                it.getInt(Publishers.PUBLISHER_ID),
+                                it.getString(Publishers.PUBLISHER_NAME)
+                        ))
                     } while (it.moveToNext())
                 }
             }
@@ -235,15 +245,18 @@ class GameDao(private val context: BggApplication) {
         }
     }
 
-    fun loadCategories(gameId: Int): LiveData<List<Pair<Int, String>>> {
+    fun loadCategories(gameId: Int): LiveData<List<GameDetailEntity>> {
         if (gameId == BggContract.INVALID_ID) return AbsentLiveData.create()
         val uri = Games.buildCategoriesUri(gameId)
         return RegisteredLiveData(context, uri) {
-            val results = arrayListOf<Pair<Int, String>>()
+            val results = arrayListOf<GameDetailEntity>()
             context.contentResolver.load(uri)?.use {
                 if (it.moveToFirst()) {
                     do {
-                        results.add(it.getInt(Categories.CATEGORY_ID) to it.getString(Categories.CATEGORY_NAME))
+                        results.add(GameDetailEntity(
+                                it.getInt(Categories.CATEGORY_ID),
+                                it.getString(Categories.CATEGORY_NAME)
+                        ))
                     } while (it.moveToNext())
                 }
             }
@@ -251,15 +264,18 @@ class GameDao(private val context: BggApplication) {
         }
     }
 
-    fun loadMechanics(gameId: Int): LiveData<List<Pair<Int, String>>> {
+    fun loadMechanics(gameId: Int): LiveData<List<GameDetailEntity>> {
         if (gameId == BggContract.INVALID_ID) return AbsentLiveData.create()
         val uri = Games.buildMechanicsUri(gameId)
         return RegisteredLiveData(context, uri) {
-            val results = arrayListOf<Pair<Int, String>>()
+            val results = arrayListOf<GameDetailEntity>()
             context.contentResolver.load(uri)?.use {
                 if (it.moveToFirst()) {
                     do {
-                        results.add(it.getInt(Mechanics.MECHANIC_ID) to it.getString(Mechanics.MECHANIC_NAME))
+                        results.add(GameDetailEntity(
+                                it.getInt(Mechanics.MECHANIC_ID),
+                                it.getString(Mechanics.MECHANIC_NAME)
+                        ))
                     } while (it.moveToNext())
                 }
             }
@@ -267,19 +283,63 @@ class GameDao(private val context: BggApplication) {
         }
     }
 
-    fun loadExpansions(gameId: Int, inbound: Boolean = false): LiveData<List<Pair<Int, String>>> {
+    fun loadExpansions(gameId: Int, inbound: Boolean = false): LiveData<List<GameExpansionsEntity>> {
         if (gameId == BggContract.INVALID_ID) return AbsentLiveData.create()
         val uri = Games.buildExpansionsUri(gameId)
         return RegisteredLiveData(context, uri) {
-            val results = arrayListOf<Pair<Int, String>>()
+            val results = arrayListOf<GameExpansionsEntity>()
             context.contentResolver.load(uri,
                     selection = "${GamesExpansions.INBOUND}=?",
                     selectionArgs = arrayOf(if (inbound) "1" else "0"))?.use {
                 if (it.moveToFirst()) {
                     do {
-                        results.add(it.getInt(GamesExpansions.EXPANSION_ID) to it.getString(GamesExpansions.EXPANSION_NAME))
+                        results.add(GameExpansionsEntity(
+                                it.getInt(GamesExpansions.EXPANSION_ID),
+                                it.getString(GamesExpansions.EXPANSION_NAME)
+                        ))
                     } while (it.moveToNext())
                 }
+            }
+            val projection = arrayOf(
+                    Collection.STATUS_OWN,
+                    Collection.STATUS_PREVIOUSLY_OWNED,
+                    Collection.STATUS_FOR_TRADE,
+                    Collection.STATUS_WANT,
+                    Collection.STATUS_WANT_TO_BUY,
+                    Collection.STATUS_WISHLIST,
+                    Collection.STATUS_WANT_TO_PLAY,
+                    Collection.STATUS_PREORDERED,
+                    Collection.STATUS_WISHLIST_PRIORITY,
+                    Collection.NUM_PLAYS,
+                    Collection.RATING,
+                    Collection.COMMENT
+            )
+            for (result in results) {
+                context.contentResolver.load(Collection.CONTENT_URI,
+                        projection = projection,
+                        selection = "collection.${Collection.GAME_ID}=?",
+                        selectionArgs = arrayOf(result.id.toString()))?.use {
+                    if (it.moveToFirst()) {
+                        do {
+                            result.apply {
+                                own = it.getBoolean(Collection.STATUS_OWN)
+                                previouslyOwned = it.getBoolean(Collection.STATUS_PREVIOUSLY_OWNED)
+                                preOrdered = it.getBoolean(Collection.STATUS_PREORDERED)
+                                forTrade = it.getBoolean(Collection.STATUS_FOR_TRADE)
+                                wantInTrade = it.getBoolean(Collection.STATUS_WANT)
+                                wantToPlay = it.getBoolean(Collection.STATUS_WANT_TO_PLAY)
+                                wantToBuy = it.getBoolean(Collection.STATUS_WANT_TO_BUY)
+                                wishList = it.getBoolean(Collection.STATUS_WISHLIST)
+                                wishListPriority = it.getIntOrNull(Collection.STATUS_WISHLIST_PRIORITY) ?: 3
+                                numberOfPlays = it.getIntOrZero(Collection.NUM_PLAYS)
+                                rating = it.getDoubleOrZero(Collection.RATING)
+                                comment = it.getString(Collection.COMMENT)
+                            }
+
+                        } while (it.moveToNext())
+                    }
+                }
+                result.id
             }
             return@RegisteredLiveData results
         }