+++ app/src/main/java/com/boardgamegeek/entities/GameDetailEntity.kt
@@ -1,0 +1,7 @@
+package com.boardgamegeek.entities
+
+data class GameDetailEntity(
+        val id: Int,
+        val name: String,
+        val description: String = ""
+)