+++ app/src/main/java/com/boardgamegeek/extensions/List.kt
@@ -6,16 +6,17 @@ fun <E> List<E>?.formatList(and: String = "&", comma: String = ","): String {
         this.isEmpty() -> return ""
         size == 1 -> return this[0].toString()
         size == 2 -> return "${this[0]} $and ${this[1]}"
-        else -> for (i in indices) {
+        else -> {
             val sb = StringBuilder()
-            sb.append(this[i])
-            if (i == size - 2) {
-                sb.append(comma).append(" ").append(and).append(" ")
-            } else if (i < size - 2) {
-                sb.append(comma).append(" ")
+            for (i in indices) {
+                sb.append(this[i])
+                if (i == size - 2) {
+                    sb.append(comma).append(" ").append(and).append(" ")
+                } else if (i < size - 2) {
+                    sb.append(comma).append(" ")
+                }
             }
             return sb.toString()
         }
     }
-    return ""
 }